pyproject.toml
[project]
name = "portfolio-chat"
version = "0.1.0"
description = "Personal RAG chatbot backend"
readme = "README.md"
requires-python = ">=3.10,<3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",

    "langchain==0.2.16",
    "langchain-core==0.2.38",
    "langchain-community==0.2.16",
    "langchain-text-splitters==0.2.4",

    "langchain-groq>=0.1.9",
    "langchain-huggingface>=0.0.3",
    "sentence-transformers>=2.7.0",
    "faiss-cpu>=1.8.0.post1"
]
